+++ src/main/java/ru/dpopkov/epmtask/model/Task.java
@@ -1,0 +1,43 @@
+package ru.dpopkov.epmtask.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.Duration;
+
+@Data
+@NoArgsConstructor
+public class Task {
+
+    private Long id;
+    private String name;
+    private String description;
+    private Project project;
+    private State state = State.OPEN;
+    private Duration time;
+    private User assignedTo;
+
+    public Task(String name, String description, Project project, Duration time) {
+        this.name = name;
+        this.description = description;
+        this.project = project;
+        this.time = time;
+    }
+
+    public void close() {
+        setState(State.CLOSED);
+    }
+
+    public boolean isOpen() {
+        return state == State.OPEN;
+    }
+
+    public boolean isClosed() {
+        return state == State.CLOSED;
+    }
+
+    public enum State {
+        OPEN,
+        CLOSED
+    }
+}